refactor(ui): move RichTextState and its tests to the root ui package
